// web/src/api/generated/models/NotificationItem.ts
export interface NotificationItem {
    /**
     * 
     * @type {AuthorSummary}
     * @memberof NotificationItem
     */
    actor?: AuthorSummary | null;
    /**
     * 
     * @type {string}
     * @memberof NotificationItem
     */
    actorId: string | null;
    /**
     * 
     * @type {Date}
     * @memberof NotificationItem
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof NotificationItem
     */
    id: string;
    /**
     * 
     * @type {NotificationKind}
     * @memberof NotificationItem
     */
    kind: NotificationKind;
    /**
     * 
     * @type {Date}
     * @memberof NotificationItem
     */
    readAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof NotificationItem
     */
    sourceEventId: string;
    /**
     * 
     * @type {ActivityTargetPresentation}
     * @memberof NotificationItem
     */
    target?: ActivityTargetPresentation | null;
    /**
     * 
     * @type {string}
     * @memberof NotificationItem
     */
    targetId: string | null;
    /**
     * 
     * @type {EngagementTarget}
     * @memberof NotificationItem
     */
    targetType: EngagementTarget | null;
}